# KV cache quantized to q8_0