--- a/llama.h
+++ b/llama.h
@@ -140,10 +140,11 @@ extern "C" {
         // context pointer passed to the progress callback
         void * progress_callback_user_data;
 
+        enum ggml_type kv_type; // the type to use for the KV cache
+
         // Keep the booleans together to avoid misalignment during copy-by-value.
         bool low_vram;   // if true, reduce VRAM usage at the cost of performance
         bool mul_mat_q;  // if true, use experimental mul_mat_q kernels
-        bool f16_kv;     // use fp16 for KV cache
         bool logits_all; // the llama_eval() call computes all logits, not just the last one
         bool vocab_only; // only load the vocabulary, no weights
         bool use_mmap;   // use mmap if possible