This PR aims to implement quantizing the KV cache as q8_0. As of right now it is a proof of concept and I would like to discuss the best way of doing a proper implementation ahead of time.

Potential benefits:

* Roughly 50% less RAM/VRAM usage for the KV cache, particularly relevant for long contexts and batched evaluation, as may become more important due to https://github.com/ggerganov/llama.cpp/pull/2926 .
* Faster evaluation due to integer SIMD instructions and less I/O.
* Less reliance on external BLAS libraries that are at this point mostly needed for floating point GEMM; as of right now the KV cache is the only remaining use case for quantized models.

The current state of this PR is that only the CPU implementation works (with some hacks). These are the performance results on my system:

| Model    | CUDA? | ngl | test       | t/s master       | t/s PR        | Speedup |
| -------- | ----- | --: | ---------- | ---------------: | ------------- | ------- |
| 7B Q4_0  | No    |   0 | pp 512     | 25.05 ± 0.04     | 25.12 ± 0.20  | 1.00    |
| 7B Q4_0  | Yes   |   0 | pp 512     | 309.63 ± 1.71    | 289.90 ± 0.60 | 0.94    |
| 7B Q4_0  | Yes   |  33 | pp 512     | 585.59 ± 2.64    | 519.90 ± 1.97 | 0.89    |
| 7B Q4_0  | No    |   0 | tg 128     | 9.99 ± 0.03      | 10.03 ± 0.03  | 1.00    |
| 7B Q4_0  | Yes   |   0 | tg 128     | 9.98 ± 0.04      | 9.97 ± 0.05   | 1.00    |
| 7B Q4_0  | Yes   |  33 | tg 128     | 89.74 ± 0.13     | 84.01 ± 0.02  | 0.94    |

The GPU is an RTX 3090. The use of CUDA up to 33 layers is possible because only at 34 layers the KV cache starts being offloaded. In its current form there seems to be a small performance regression. The Perplexity on wikitext-2 changed from 5.9679 +/- 0.03352 to 5.9668 +/- 0.03349 which is actually a very small improvement, likely due to randomness (calculated with 33 layers offloaded).

The implementation for the K component of the KV cache was fairly straightforward: the values copied to the K component always have a row size of `n_embd_head` which is a multiple of 32 (q8_0 block size). So you can essentially just change the datatype of the K component to q8_0 and the preexisting implementations for copying and matrix multiplication will work out-of-the-box.

The implementation for the V component was more tricky because the row size is equal to the number of processed tokens which then need to be copied to an arbitrary location in a q8_0 block. I did the following things to make it work:

* Extend `ggml_compute_forward_dup_f32` by an extra case that so far was not implemented: copying non-contiguous f32 values to non-contiguous q8_0 values. My solution was to fill a zero-initialized temporary buffer behind the q8_0 block with the unquantized values. Then re-create the q8_0 block from those when either the buffer contains enough values for an entire q8_0 block or when there are no more new unquantized values to copy for a row. This requires an additional 128 bytes per V component row (hacked by just increasing the size by 128 elements).
* The aforementioned extra case needs to know the starting position inside the q8_0 block in order to determine the location of the temporary buffer and when it should be converted to q8_0. I hacked this by writing this information to the `op_params` of the corresponding tensor in `llm_build_llama`.
* The matrix multiplication code for q8_0 expects the row size to be a multiple of 32 (q8_0 block size). For the V component of the KV cache this can be done by just taking a larger slice from it. The hidden state needs to be padded with extra values. I hacked this by creating a new tensor that has the padded size, then creating a view of that tensor, and then copying the hidden state to said view.

I plan to do a proper implementation of this PR in the following way:

* Modify `ggml_view` so that the position inside a block is written to `op_params` if the underlying tensor is quantized.
* Add a `GGML_OP` like `GGML_OP_PAD` that pads a tensor with a constant value. Alternatively the code for quantizing the hidden state could be modified which would probably have better performance. But this would also be more complicated and require changes across a much wider range of hardware (to some of which I don't have access for testing). For the CUDA implementation only very minimal changes would be necessary to avoid padding the hidden state in an extra `GGML_OP`.
* I will implement q8_0 KV cache quantization for CPU and CUDA.
* I will replace the `--memory-f32` CLI argument with a new one that lets you choose between q8_0, f16, and f32 for the KV cache type.

If there are issues with the way I plan to do the implementation, please let me know ahead of time.