@JohannesGaessler 2023-09-02
>Add a GGML_OP like GGML_OP_PAD that pads a tensor with a constant value. Alternatively the code for quantizing the hidden state could be modified which would probably have better performance.

On second thought, maybe not. The order of tensors would be matrix multiplication -> scale -> mask -> softmax -> pad -> matrix multiplication. For optimal performance you should be fusing those tensors anyways and adding one more unary tensor to the pipeline would not make a difference.