@JohannesGaessler 2023-09-02
>Why do you add 128 again?

I'm adding 128 so each row has some extra space to temporarily store the unquantized values as the context fills up.
It's to avoid precision loss when only part of a q8_0 block has been filled.