@KerfuffleV2 2023-09-02
This seems to work on ROCM (RX 6600) but weirdly uses considerably more memory than `master` for the same context size. Generation speed was about the same (perplexity not tested). This was a 7B LLaMA1 Q4_K model with `-ngl 33`. For master, it used ~`4,500 MiB` VRAM. This pull ~`6,500 MiB`.