@JohannesGaessler 2023-09-02
>Also, probably better to utilize the wdata buffer that each op uses for writing work data

From what I can tell `wdata` would not be suitable because I need persistence across multiple eval calls.

>This seems to work on ROCM (RX 6600) but weirdly uses considerably more memory than master for the same context size.

I think this is due to the hack that I'm currently using to pad one of the tensors. It should not be an issue in the final version.