@kiratp 2023-09-02
> Less reliance on external BLAS libraries that are at this point mostly needed for floating point GEMM; as of right now the KV cache is the only remaining use case for quantized models.


Unless the plan is to call the new AMX (and VNNI) instructions manually, not using MKL (would be slower on any newer (post Sapphire Rapids) Intel CPUs.