@JohannesGaessler 2023-09-04
I added new functions like `ggml_view_blck_1d` that allow you to set the starting position in a block (I think this is preferable over changing the interface of e.g. `ggml_view_1d` and adding an extra parameter).

I replaced the padding hack with an extension to the matrix multiplication ggml code. The solution that I came up with is to quantize the last block of the hidden state without any SIMD instruction if it is incomplete. The rest of the tensor is still quantiized using SIMD instructions. This should keep implementation/maintenance overhead low with negligible impact on performance.

I pushed CUDA implementations that produce correct results but the performance is bad.