@Cebtenzzre 2023-09-05
test-opt and test-grad0 are failing an assertion:
```
GGML_ASSERT: ggml.c:11303: from_float_to_vec_dot != NULL
```