@JohannesGaessler 2023-09-05
Alright, this PR should now be feature complete. I added a new CLI argument `--kv-type` that lets users set the type as which the KV cache is stored. By default q8_0 is used since the memory savings are quite substantial: for 4096 context with 7b q4_0 the VRAM usage goes down from 6478 MiB to 5536 MiB (942 MiB reduction). The token generation performance seems to be slightly slower, prompt processing on average a little faster:

| model   | GPU      | test       | t/s master      | t/s PR           | Speedup |
| -----   | RTX 3090 | ---------- | --------------- | ---------------: | ------- |
| 7b q4_0 | RTX 3090 | pp 512     | 2187.37 ± 34.18 | 2303.08 ± 24.19  |    1.05 |
| 7b q4_0 | RTX 3090 | tg 128     | 141.11 ± 0.30   | 140.19 ± 0.31    |    0.99 |
| 7b q4_0 | P40      | pp 512     | 875.56 ± 4.72   | 870.33 ± 2.09    |    0.99 |
| 7b q4_0 | P40      | tg 128     | 59.45 ± 0.07    | 57.23 ± 0.01     |    0.96 |
| 7b q4_0 | RX 6800  | pp 512     | 1061.05 ± 6.51  | 1159.28 ± 7.16   |    1.09 |
| 7b q4_0 | RX 6800  | tg 128     | 77.89 ± 0.02    | 72.95 ± 0.01     |    0.94 |

The perplexity changes as follows:

| model     | PPL master | PPL PR | PPL diff |
|-----------|------------|--------|----------|
| 7b f16    |     5.7962 | 5.7989 |   2.7e-3 |
| 7b q4_0   |     5.9670 | 5.9658 |  -1.2e-3 |
| 7b q4_1   |     6.0035 | 6.0038 |     3e-4 |
| 7b q5_0   |     5.8292 | 5.8327 |   3.5e-3 |
| 7b q5_1   |     5.8538 | 5.8560 |   2.2e-3 |
| 7b q8_0   |     5.8014 | 5.8018 |     4e-4 |
| 7b q2_k   |     6.4466 | 6.4514 |   4.8e-3 |
| 7b q3_k_s |     6.2950 | 6.2955 |     5e-4 |
| 7b q3_k_m |     6.0264 | 6.0297 |   3.3e-3 |
| 7b q3_k_l |     5.9854 | 5.9911 |   5.7e-3 |
| 7b q4_k_s |     5.8891 | 5.8934 |   4.3e-3 |
| 7b q4_k_m |     5.8798 | 5.8840 |   4.2e-3 |
| 7b q5_k_s |     5.8215 | 5.8260 |   4.5e-3 |
| 7b q5_k_m |     5.8292 | 5.8332 |     4e-3 |
| 7b q6_k   |     5.8098 | 5.8141 |   4.3e-3 |

The increase in perplexity is on the same order as the perplexity differences between f16 and q8_0 or q8_0 and q6_K.