@Cebtenzzre 2023-09-05
`make LLAMA_CUBLAS=1` is failing on my computer:
```
ggml-cuda.cu(2107): error: the size of an array must be greater than zero
      __attribute__((shared)) int tile_x_qs[mmq_y * (32) + mmq_y];
                                            ^
          detected during:
            instantiation of "void allocate_tiles_q4_0<mmq_y>(int **, half2 **, int **, int **) [with mmq_y=-1]" at line 3483
            instantiation of "void mul_mat_q4_0<need_check>(const void *, const void *, float *, int, int, int, int, int, int, int, int) [with need_check=false]" at line 4662
```