@KerfuffleV2 2023-09-05
Running perplexity with ROCM, I just get NANs when `-kvt q8_0`. It doesn't seem to make a difference whether `-ngl 0` or offloading the whole model.

```plaintext
ggml_init_cublas: found 1 ROCm devices:
  Device 0: AMD Radeon RX 6600, compute capability 10.3
[...]
llm_load_tensors: ggml ctx size =    0.08 MB
llm_load_tensors: using ROCm for GPU acceleration
llm_load_tensors: mem required  =  103.84 MB (+   97.12 MB per state)
llm_load_tensors: offloading 26 repeating layers to GPU
llm_load_tensors: offloading non-repeating layers to GPU
llm_load_tensors: offloading v cache to GPU
llm_load_tensors: offloading k cache to GPU
llm_load_tensors: offloaded 29/29 layers to GPU
llm_load_tensors: VRAM used: 3466 MB
[...]
llama_new_context_with_model: kv self size  =   97.12 MB
llama_new_context_with_model: compute buffer total size =   70.22 MB
llama_new_context_with_model: VRAM scratch buffer: 68.75 MB
system_info: n_threads = 5 / 16 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | 
perplexity: tokenizing the input ..
perplexity: tokenization took 759.39 ms
perplexity: calculating perplexity over 616 chunks, batch_size=512
perplexity: 0.59 seconds per pass - ETA 6.10 minutes
[1]nan,[2]nan,[3]nan,[4]nan,[5]nan,[6]nan,[7]nan,[8]nan,[9]nan,[10]nan,[11]nan,[12]nan,[13]nan
```

Token generation seemed to work okay.