@JohannesGaessler 2023-09-05
Do you still get NaN with `-nommq`?