@KerfuffleV2 2023-09-05
> Do you still get NaN with `-nommq`?

Yes, I do.

```plaintext
perplexity: 1.27 seconds per pass - ETA 13.03 minutes
[1]nan,[2]nan,[3]nan
```

I get them slower! (Not that this information is likely to be too helpful.)