@Cebtenzzre 2023-09-05
I get an assertion failure if I set my GTX 970 as the main GPU with `-mg 1`.
Full command: `./main -ngl 100 -t 1 -m llama-7b.q4_0.gguf -n 128 --ignore-eos -mg 1`
```
GGML_ASSERT: ggml-cuda.cu:5782: buffers_contiguous || ne02 == 1
```