@JohannesGaessler 2023-09-05
I've pushed a fix for incorrect results with compute capabilities < 6.1. The results should now be correct but the performance will be bad.