@KerfuffleV2 2023-09-05
Tested at e0d5c0bce812353631c736396d047e282ad72eff - no difference as far as I can see. With `-ngl 0`, with `-nommq`, both. All NaNs. Perhaps your compute level checking isn't working correctly for ROCM? The speed seems the same as before.