@sozforex 2023-09-11
I can replicate what Dampfinchen have described.
On master in all cases described below things work ok.

### PR: -c 64 -n 128 -ngl 35  (35/35 layers offloaded, same with 34/35)
```
./main -m ../models/gguf/codellama-7b.Q4_K_M.gguf -p '// Quick-sort implementation in C (4 spaces indentation + detailed comments) and sample usage:\n\n#include' -e -s 8 --top_k 1 --temp 0.
1 --top-p 0.95 -t 8 -c 128 -n 256 -ngl 35

...

#include <stdio.h>
#include <stdlib.h>

// Function prototypes
void quick_sort(int *, int);
void swap(int *, intGGML_ASSERT: /home/xerofzos/code/ai/llama.cpp/ggml-cuda.cu:5127: i_blck_0 + ne00 < QK8_0
zsh: IOT instruction (core dumped)  HSA_OVERRIDE_GFX_VERSION=10.3.0 HIP_VISIBLE_DEVICES=0 build_kv/bin/main -m  -
```

### PR: -c 64 -n 128 -ngl 33  (33/35 layers offloaded, same with 1/35)

```
#include <stdio.h>
#include <stdlib.h>

// Function prototypes
void quick_sort(int *, int);
void swap(int *, int▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅
llama_print_timings:        load time =   916,58 ms
llama_print_timings:      sample time =    41,79 ms /   128 runs   (    0,33 ms per token,  3063,08 tokens per second)
llama_print_timings: prompt eval time =   667,52 ms /   124 tokens (    5,38 ms per token,   185,76 tokens per second)
llama_print_timings:        eval time =  2647,56 ms /   124 runs   (   21,35 ms per token,    46,84 tokens per second)
llama_print_timings:       total time =  3372,78 ms
Log end
```

### PR: -c 64 -n 128 -ngl 0  (0/35 layers offloaded)

```
#include <stdio.h>
#include <stdlib.h>

// Function prototypes
void quick_sort(int *, int);
void swap(int *, int▅eaLYصindeneaeaINCTслу Agrompeaeaeaeaeaeaeaeaeaeaeaeaeaeaeaeaeaeaeaea▅authenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticauthenticezcredauthenticoucauthenticauthenticauthenticauthenticauthentic▅aven resident rolling resident resident resident resident residentanne norm resident resident resident resident resident resident resident resident resident resident resident resident resident resident resident
llama_print_timings:        load time =   181,56 ms
llama_print_timings:      sample time =    41,51 ms /   128 runs   (    0,32 ms per token,  3083,30 tokens per second)
llama_print_timings: prompt eval time =  3344,18 ms /   124 tokens (   26,97 ms per token,    37,08 tokens per second)
llama_print_timings:        eval time = 10720,33 ms /   124 runs   (   86,45 ms per token,    11,57 tokens per second)
llama_print_timings:       total time = 14124,83 ms
Log end
```